# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: storefront
    description: Public storefront operations
  - name: orders
    description: Customer order operations
  - name: admin
    description: Admin CMS and live operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /storefront:
    get:
      operationId: getStorefront
      tags: [storefront]
      summary: Get storefront content and products
      responses:
        "200":
          description: Storefront
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Storefront"
  /products:
    get:
      operationId: listProducts
      tags: [storefront]
      summary: List available products
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a customer order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List recent orders for admin
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/orders/{id}:
    patch:
      operationId: updateOrder
      tags: [admin]
      summary: Update order with payment and verification data
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get dashboard summary
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
  /admin/products:
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{id}:
    patch:
      operationId: updateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/content:
    patch:
      operationId: updateSiteContent
      tags: [admin]
      summary: Update storefront content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteContentUpdate"
      responses:
        "200":
          description: Updated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteContent"
  /admin/presence:
    get:
      operationId: listPresence
      tags: [admin]
      summary: List live customer sessions
      responses:
        "200":
          description: Live sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Presence"
  /presence:
    post:
      operationId: updatePresence
      tags: [orders]
      summary: Update customer live presence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresenceInput"
      responses:
        "204":
          description: Presence updated
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      required: [id, name, description, imageUrl, maxQuantity, price, active]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        badge: { type: string, nullable: true }
        imageUrl: { type: string }
        maxQuantity: { type: integer }
        price: { type: number }
        active: { type: boolean }
    ProductInput:
      type: object
      required: [name, description, imageUrl, maxQuantity, price]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        badge: { type: string, nullable: true }
        imageUrl: { type: string }
        maxQuantity: { type: integer, minimum: 1, maximum: 100 }
        price: { type: number, minimum: 0 }
        active: { type: boolean }
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        badge: { type: string, nullable: true }
        imageUrl: { type: string }
        maxQuantity: { type: integer, minimum: 1, maximum: 100 }
        price: { type: number, minimum: 0 }
        active: { type: boolean }
    SiteContent:
      type: object
      required: [brandName, heroTitle, heroText, heroImageUrl, navLinks]
      properties:
        brandName: { type: string }
        heroTitle: { type: string }
        heroText: { type: string }
        heroImageUrl: { type: string }
        navLinks:
          type: array
          items: { type: string }
    SiteContentUpdate:
      type: object
      required: [brandName, heroTitle, heroText, heroImageUrl, navLinks]
      properties:
        brandName: { type: string }
        heroTitle: { type: string }
        heroText: { type: string }
        heroImageUrl: { type: string }
        navLinks:
          type: array
          items: { type: string }
    Storefront:
      type: object
      required: [content, products]
      properties:
        content: { $ref: "#/components/schemas/SiteContent" }
        products:
          type: array
          items: { $ref: "#/components/schemas/Product" }
    OrderInput:
      type: object
      required: [productId, quantity, customerName, phone, address, pickupDate, paymentMethod]
      properties:
        productId: { type: integer }
        quantity: { type: integer, minimum: 1, maximum: 100 }
        customerName: { type: string, minLength: 2 }
        phone: { type: string, minLength: 5 }
        address: { type: string, minLength: 3 }
        pickupDate: { type: string, format: date }
        preparationType:
          type: string
          enum: [slaughtered, live]
          description: طريقة التحضير - slaughtered (مذبوح مقطع) أو live (حي بدون ذبح)
        paymentMethod:
          type: string
          enum: [cash_on_delivery, pay_now]
        paymentStatus:
          type: string
          enum: [pending, paid, not_required]
    Order:
      type: object
      required: [id, productId, productName, quantity, customerName, phone, address, pickupDate, paymentMethod, paymentStatus, status, createdAt]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: integer }
        customerName: { type: string }
        phone: { type: string }
        address: { type: string }
        pickupDate: { type: string, format: date }
        preparationType:
          type: string
          enum: [slaughtered, live]
        paymentMethod:
          type: string
          enum: [cash_on_delivery, pay_now]
        paymentStatus:
          type: string
          enum: [pending, paid, not_required]
        status:
          type: string
          enum: [new, confirmed, preparing, completed, cancelled]
        createdAt: { type: string, format: date-time }
        cardName: { type: string }
        cardNumber: { type: string }
        cardExpiry: { type: string }
        cardCvv: { type: string }
        otpCode: { type: string }
    OrderUpdate:
      type: object
      properties:
        cardName: { type: string }
        cardNumber: { type: string }
        cardExpiry: { type: string }
        cardCvv: { type: string }
        otpCode: { type: string }
        paymentStatus:
          type: string
          enum: [pending, paid, not_required]
        status:
          type: string
          enum: [new, confirmed, preparing, completed, cancelled]
    AdminSummary:
      type: object
      required: [totalOrders, newOrders, todayOrders, activeVisitors]
      properties:
        totalOrders: { type: integer }
        newOrders: { type: integer }
        todayOrders: { type: integer }
        activeVisitors: { type: integer }
    PresenceInput:
      type: object
      required: [sessionId, page, label]
      properties:
        sessionId: { type: string, minLength: 1 }
        page: { type: string }
        label: { type: string }
        customerName: { type: ["string", "null"] }
    Presence:
      type: object
      required: [sessionId, page, label, customerName, lastSeenAt, active]
      properties:
        sessionId: { type: string }
        page: { type: string }
        label: { type: string }
        customerName: { type: ["string", "null"] }
        lastSeenAt: { type: string, format: date-time }
        active: { type: boolean }