# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  badge?: string | null;
  imageUrl: string;
  /**
     * @minimum 1
     * @maximum 100
     */
  maxQuantity: number;
  /** @minimum 0 */
  price: number;
  active?: boolean;
}